requirements.txt
#this is the "requirements.txt" file...

python-dotenv

requests 

pandas

plotly

sendgrid==6.5.0

pytest

flask

gunicorn